feat: PLGR-53 move GameResultCard to organisms folder
