+++ packages/ui/molecules/index.ts
@@ -1,4 +1,3 @@
 export { LabeledInput } from './labeledInput';
 export { ScoringButton } from './scoring-button';
-export { GameResultCard } from './gameResultCard';
 export { GameResultCardScoreBadge } from './gameResultCardScoreBadge';
