+++ packages/ui/organisms/gameResultCard/index.ts
