+++ packages/ui/organisms/index.ts
@@ -1,5 +1,6 @@
 export { AuthForm } from './authForm';
-export { SubNavigation } from "./subNavigation";
+export { SubNavigation } from './subNavigation';
 export { AuthBanner } from './authBanner';
 export { Navigation } from './navigation';
 export { GameCard } from './gameCard';
+export { GameResultCard } from './gameResultCard';
